refactor: rename polylines to editable polylines
