+++ src/components/pages/RouteViewer/index.tsx
@@ -15,7 +15,7 @@ import {
 import FocusedMarker from "../../organisms/FocusedMarker";
 import { HEADER_HEIGHT_PX } from "../../organisms/Header";
 import Markers from "../../organisms/EditableMarkers";
-import Polylines from "../../organisms/Polylines";
+import Polylines from "../../organisms/EditablePolylines";
 import RouteEditController from "../../organisms/RouteEditController";
 import SignInRequiredTemplate from "../../organisms/SignInRequiredTemplate";
 import styles from "./style.module.css";
